Add files via upload
License plate recognition is one of the most exciting and stimulating research topics in recent 
years. The license plates of each country are shown to be of different shapes and sizes, as well 
as different colors. In India, utmost mutual vehicle registering plates used yellow/white as the 
plate color and black as the letter color. This article used for the development of automatic 
license recognition (ANPR) by using Python, OpenCV and the EasyOCR library. ANPR 
technology used to identifies and extracts vehicle license from photos or videos. OpenCV used 
for image pre-processing and control, while EasyOCR stands for optical character recognition 
(OCR) to recognize license plates. This article includes following steps to create an ANPR 
system, including image creation, character classification, and OCR. Issues related to strategic 
implementation and resource development are also discussed. It also shows an easy way to 
segment all the alphabets and figures used in license plates. The first is to locate the license 
plate, and the second is to slice all the words and alphabets to recognize each no individually.